Reorganise, again
Basically I'm turning this into a Laravel application
