+++ app/server.js
@@ -4,8 +4,8 @@ const Hapi = require('@hapi/hapi')
 const server = Hapi.server({ host: 'localhost', port: 3000 })
 
 const routes = [
-  ...require('./routes/home'),
-  ...require('./routes/tests')
+  ...require('../routes/home'),
+  ...require('../routes/tests')
 ]
 
 routes.forEach(route => server.route(route))
