+++ test/models/QueuedTest.test.js
@@ -4,8 +4,8 @@ const Lab = require('@hapi/lab')
 const { expect } = require('@hapi/code')
 const { beforeEach, describe, it } = exports.lab = Lab.script()
 
-const QueuedTest = require('./../../lib/models/QueuedTest')
-const status = require('./../../lib/constants/testStatus')
+const QueuedTest = require('../../app/models/QueuedTest')
+const status = require('../../app/constants/testStatus')
 
 describe('QueuedTest', () => {
   let model
