+++ test/tests/read.test.js
@@ -3,7 +3,7 @@
 const Lab = require('@hapi/lab')
 const { expect } = require('@hapi/code')
 const { afterEach, beforeEach, describe, it } = exports.lab = Lab.script()
-const { init } = require('../../lib/server')
+const { init } = require('../../app/server')
 
 describe('GET /tests/{id}', () => {
   let server
